requirements.txt
pinecone
python-telegram-bot>=20.0
gradio